fix(types): widen timeout annotations to float for experiment_timeout_seconds
Mypy flagged three call sites where the new
study_execution.experiment_timeout_seconds: float field met previously
int-typed parameters:
  - DockerRunner.__init__(timeout: int | None) at infra/docker_runner.py
  - _collect_result(timeout: int) at study/runner.py

The underlying stdlib APIs (subprocess.wait, Popen.wait) accept float
natively, so the prior int annotation was an artificial narrowing rather
than a real constraint. Widen both to float (int is a structural subtype,
so existing int-literal callers continue to type-check).

+++ src/llenergymeasure/study/runner.py
@@ -356,7 +356,7 @@ def _collect_result(
     p: Any,  # multiprocessing.Process
     parent_conn: Any,  # multiprocessing.Connection (parent end)
     config: ExperimentConfig,
-    timeout: int,
+    timeout: float,
     pipe_payload: Any = _UNSET,
 ) -> Any:
     """Inspect process outcome and return either a result or a failure dict.
